migration and sweet alert

+++ resources/assets/js/components/SalesCategory/ShowSalesCategory.vue
@@ -6,7 +6,8 @@
             </h2>
             <ul class="header-dropdown m-r-0">
                 <li>
-                    <router-link :to="{ name: 'CreateSalesCategory' }" title="Add New Category" >
+                    
+                    <router-link :to="{ name: 'CreateSalesCategory' }" data-toggle="tooltip" data-placement="top" title="" data-original-title="ADD NEW CATEGORY" >
                         <i class="material-icons">add_circle_outline</i>
                     </router-link>
                 </li>
@@ -18,7 +19,6 @@
             </ul>
         </div>
         
-        <span class="waves-effect" v-if="message.message" data-placement-from="bottom" data-placement-align="center" data-animate-enter="" data-animate-exit="" data-color-name="bg-green" data-notify="message">{{message.message}}</span>
         <div class="body table-responsive">
 
             <table class="table">
@@ -37,7 +37,10 @@
                         <td>{{ category.sales_category_name }}</td>
                         <td>{{category.commition}}</td>
                         <td>{{ category.salary }}</td>
-                        <td><router-link :to="{name: 'EditSalesCategory', params: { id: category.id }}" class="btn btn-primary">Edit</router-link> <button class="btn btn-danger" v-on:click="deleteItem(category.id)">Delete</button></td>
+                        <td>
+                            <router-link :to="{name: 'EditSalesCategory', params: { id: category.id }}" class="btn btn-primary">Edit</router-link> 
+                            <button class="btn btn-danger waves-effect " v-on:click="confirmDelete(category.id)" >Delete</button>
+                        </td>
                     </tr>
                     
                 </tbody>
@@ -66,13 +69,28 @@
                   this.categories = response.data;
               });
             },
-            deleteItem(id)
-            {
-              let uri = `/admin/api/sales-category/${id}`;
-              this.axios.delete(uri).then((response) =>{
-                  this.message = response.data;
-                  this.fetchCategory();
-              });
+            confirmDelete(id){
+                this.$swal({
+                title: 'Are you sure?',
+                text: 'You can\'t revert your action',
+                type: 'warning',
+                showCancelButton: true,
+                cancelButtonText: 'No, Keep it!',
+                confirmButtonColor: "#DD6B55",
+                confirmButtonText: "Yes, delete it!",
+                showCloseButton: true,
+                showLoaderOnConfirm: true
+                }).then((result) => {
+                    if(result.value) {
+                        let uri = `/admin/api/sales-category/${id}`;
+                        this.axios.delete(uri).then((response) =>{
+                            this.message = response.data;
+                            this.$swal('Deleted', this.message.message, 'success')
+                            this.fetchCategory();
+                        });
+                        
+                    } 
+                })
             }
         }
     }
